Refs #2, moved physics object to its own file
